change: Mark gethostname tests as xfail if run locally (#220)
Running the gethostname tests requires root permissions in order to
create the /opt/ml directory. This change marks the gethostname tests
as expected to fail if not run on CodeBuild, since it won't be able
to create the /opt/ml directory.

This change checks for an environment variable set in the CodeBuild
image Dockerfile to determine whether to mark the unit test as
xfail.

This will only run the test if the "IS_CODEBUILD_IMAGE" environment
variable is assigned the string "true".