bug-fix: Typo in the execution-parameters routing rule (#239)
* feature: adding encoder to serialize in recordio-protobuf format

- Style fix per review comments

This change allows user to specify a method be used for responding
to the execution_parameters endpoint call. If the user does not specify
the function, execution_parameters endpoint will not be available.
`execution_parameters` is an optional endpoint used by SageMaker
Trasnform Job.

For more details:
https://docs.aws.amazon.com/sagemaker/latest/dg/your-algorithms-batch-code.html#your-algorithms-batch-code-how-containe-serves-requests

* feature: adding encoder to serialize in recordio-protobuf format

- Style fix per review comments

This change allows user to specify a method be used for responding
to the execution_parameters endpoint call. If the user does not specify
the function, execution_parameters endpoint will not be available.
`execution_parameters` is an optional endpoint used by SageMaker
Trasnform Job.

For more details:
https://docs.aws.amazon.com/sagemaker/latest/dg/your-algorithms-batch-code.html#your-algorithms-batch-code-how-containe-serves-requests

* bug-fix: typo in the execution-parameters url rule

* bug-fix: updated tests to use execution-parameters instead of execution_parameters